Add discover_services boolean keyword to Client init
Client._discover_services() can be an expensive operation for a data
center, but it was run on Client.__init__ even if it wasn't used.  This
keyword bypasses the service discover/query, and instead assumes the
default services at the data center.  This is helpful for distributed
applications that instantiate many clients in seperate Python processes.